+++ backend/infrahub/message_bus/operations/requests/__init__.py
@@ -1,12 +1,10 @@
 from . import (
-    artifact_definition,
     generator_definition,
     proposed_change,
     repository,
 )
 
 __all__ = [
-    "artifact_definition",
     "generator_definition",
     "proposed_change",
     "repository",
